added the removeurlparam function
