+++ src/js/handlers/listners/logout.js
@@ -1,7 +1,9 @@
 import { load } from "../../storage/_index.js";
+import { removeUrlParameter } from "../../deployment/removeUrlParam.js";
 
 export function listenForLogout(btn) {
   btn.addEventListener("click", () => {
+    removeUrlParameter("key");
     localStorage.clear();
     const token = load("token");
